fix pose details
